Fix warnings
This change fixes Cosmopolitan so it has fewer opinions about compiler
warnings. The whole repository had to be cleaned up to be buildable in
-Werror -Wall mode. This lets us benefit from things like strict const
checking. Some actual bugs might have been caught too.

+++ libc/intrin/ffs.c
@@ -16,6 +16,7 @@
 │ TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR             │
 │ PERFORMANCE OF THIS SOFTWARE.                                                │
 ╚─────────────────────────────────────────────────────────────────────────────*/
+#include "libc/nexgen32e/ffs.h"
 
 /**
  * Finds lowest set bit in word.
@@ -27,7 +28,7 @@ int ffs(int x) {
 /**
  * Finds lowest set bit in word.
  */
-long ffsl(long x) {
+int ffsl(long x) {
   return __builtin_ffsl(x);
 }
